requirements.txt
streamlit>=1.32.0
datasets
Pillow
numpy
requests 